extract Papertrail functions to own file
