+++ api/models/service.go
@@ -4,11 +4,9 @@ import (
 	"fmt"
 	"os"
 	"sort"
-	"strings"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/cloudformation"
-	"github.com/aws/aws-sdk-go/service/kinesis"
 )
 
 type Service struct {
@@ -90,202 +88,6 @@ func (s *Service) Create() error {
 	}
 
 	_, err = CloudFormation().CreateStack(req)
-
-	return err
-}
-
-func (s *Service) CreatePapertrail() error {
-	if s.URL == "" {
-		return fmt.Errorf("Papertrail URL is required")
-	}
-
-	input := struct {
-		ARNs []string
-	}{
-		[]string{},
-	}
-
-	formation, err := buildTemplate(fmt.Sprintf("service/%s", s.Type), "service", input)
-
-	if err != nil {
-		return err
-	}
-
-	if err != nil {
-		return err
-	}
-
-	params := map[string]string{
-		"Url": s.URL,
-	}
-
-	tags := map[string]string{
-		"System":  "convox",
-		"Type":    "service",
-		"Service": s.Type,
-	}
-
-	req := &cloudformation.CreateStackInput{
-		Capabilities: []*string{aws.String("CAPABILITY_IAM")},
-		StackName:    aws.String(s.Name),
-		TemplateBody: aws.String(formation),
-	}
-
-	for key, value := range params {
-		req.Parameters = append(req.Parameters, &cloudformation.Parameter{ParameterKey: aws.String(key), ParameterValue: aws.String(value)})
-	}
-
-	for key, value := range tags {
-		req.Tags = append(req.Tags, &cloudformation.Tag{Key: aws.String(key), Value: aws.String(value)})
-	}
-
-	_, err = CloudFormation().CreateStack(req)
-
-	return err
-}
-
-func (s *Service) LinkPapertrail(app App) error {
-	// get full Kinesis ARN for app
-	req, err := Kinesis().DescribeStream(&kinesis.DescribeStreamInput{
-		StreamName: aws.String(app.Outputs["Kinesis"]),
-	})
-
-	arn := *req.StreamDescription.StreamARN
-
-	if err != nil {
-		return err
-	}
-
-	// get all ARNs linked to service
-	arns := []string{}
-
-	a := s.Parameters["Arns"]
-	if a != "" {
-		arns = strings.Split(a, ",")
-	}
-
-	// append app ARN if not already linked
-	for _, a := range arns {
-		if a == arn {
-			return fmt.Errorf("app %s is already linked", app.Name)
-		}
-	}
-
-	arns = append(arns, arn)
-
-	// build Name -> ARN map for CloudFormation template EventSourceMapping resources
-	m := map[string]string{}
-
-	for _, v := range arns {
-		// e.g. arn:aws:kinesis:us-east-1:901416387788:stream/convox-Kinesis-L6MUKT1VH451 -> convox
-		a := strings.Split(strings.Split(v, "/")[1], "-")[0]
-		m[a] = v
-	}
-
-	input := struct {
-		ARNs map[string]string
-	}{
-		m,
-	}
-
-	formation, err := buildTemplate(fmt.Sprintf("service/%s", s.Type), "service", input)
-
-	if err != nil {
-		return err
-	}
-
-	// Update stack with all linked ARNs and EventSourceMappings
-	_, err = CloudFormation().UpdateStack(&cloudformation.UpdateStackInput{
-		StackName:    aws.String(s.Name),
-		Capabilities: []*string{aws.String("CAPABILITY_IAM")},
-		Parameters: []*cloudformation.Parameter{
-			&cloudformation.Parameter{
-				ParameterKey:   aws.String("Arns"),
-				ParameterValue: aws.String(strings.Join(arns, ",")),
-			},
-			&cloudformation.Parameter{
-				ParameterKey:   aws.String("Url"),
-				ParameterValue: aws.String(s.Parameters["Url"]),
-			},
-		},
-		TemplateBody: aws.String(formation),
-	})
-
-	return err
-}
-
-func (s *Service) UnlinkPapertrail(app App) error {
-	// get full Kinesis ARN for app
-	req, err := Kinesis().DescribeStream(&kinesis.DescribeStreamInput{
-		StreamName: aws.String(app.Outputs["Kinesis"]),
-	})
-
-	arn := *req.StreamDescription.StreamARN
-
-	if err != nil {
-		return err
-	}
-
-	// get all ARNs linked to service
-	arns := []string{}
-
-	a := s.Parameters["Arns"]
-
-	if a != "" {
-		arns = strings.Split(a, ",")
-	}
-
-	if !strings.Contains(a, arn) {
-		return fmt.Errorf("app %s is not linked", app.Name)
-	}
-
-	// append ARNs except for app to unlink
-	newArns := []string{}
-	for _, a := range arns {
-		if a != arn {
-			newArns = append(newArns, a)
-		}
-	}
-
-	arns = newArns
-
-	// build Name -> ARN map for CloudFormation template EventSourceMapping resources
-	m := map[string]string{}
-
-	for _, v := range arns {
-		// e.g. arn:aws:kinesis:us-east-1:901416387788:stream/convox-Kinesis-L6MUKT1VH451 -> convox
-		a := strings.Split(strings.Split(v, "/")[1], "-")[0]
-		m[a] = v
-	}
-
-	input := struct {
-		ARNs map[string]string
-	}{
-		m,
-	}
-
-	formation, err := buildTemplate(fmt.Sprintf("service/%s", s.Type), "service", input)
-
-	if err != nil {
-		return err
-	}
-
-	// Update stack with all linked ARNs and EventSourceMappings
-	_, err = CloudFormation().UpdateStack(&cloudformation.UpdateStackInput{
-		StackName:    aws.String(s.Name),
-		Capabilities: []*string{aws.String("CAPABILITY_IAM")},
-		Parameters: []*cloudformation.Parameter{
-			&cloudformation.Parameter{
-				ParameterKey:   aws.String("Arns"),
-				ParameterValue: aws.String(strings.Join(arns, ",")),
-			},
-			&cloudformation.Parameter{
-				ParameterKey:   aws.String("Url"),
-				ParameterValue: aws.String(s.Parameters["Url"]),
-			},
-		},
-		TemplateBody: aws.String(formation),
-	})
 
 	return err
 }
